feat(simulator): support handle ur logic in simulator
