+++ src/ui/gui_model/gui_model.c
@@ -1139,14 +1139,7 @@ static int32_t ModeControlQrDecode(const void *inData, uint32_t inDataLen)
         PubValueMsg(QRDECODE_MSG_STOP, 0);
     }
 #else
-    static uint8_t urRet = 0;
-#ifndef BTC_ONLY
-    UrViewType_t urViewType = { CardanoTx, CardanoSignRequest };
-#else
-    UrViewType_t urViewType = { BtcNativeSegwitTx, CryptoPSBT };
-    read_qrcode(&urViewType);
-#endif
-    GuiEmitSignal(SIG_QRCODE_VIEW_SCAN_PASS, &urViewType, sizeof(urViewType));
+    read_qrcode();
 #endif
     SetLockScreen(enable);
     return SUCCESS_CODE;
